requirements.txt
# Web Framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# Environment & Config
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1

# Image & PDF Processing
pillow==10.3.0
pymupdf==1.24.3

# OCR
easyocr==1.7.1

# AI
groq==0.8.0

# File handling
aiofiles==23.2.1